+++ frontend/src/app/products/page.tsx
@@ -1,177 +1,26 @@
 import ClientPage from "./ClientPage";
 import { Product } from "../types"; // Product 타입 임포트
 
+interface ProductResponse {
+  statusCode: number;
+  message: string;
+  data: Product[];
+}
+
 export default async function Page() {
-  const products: Product[] = [
-    {
-      id: 1,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 2,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 3,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 4,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 5,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 6,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 7,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 8,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 9,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 10,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 11,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 12,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 13,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 14,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 15,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 16,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 17,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 18,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 19,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 20,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 21,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-    {
-      id: 22,
-      name: "블루투스 키보드",
-      price: 35000,
-      description:
-        "편안한 타이핑을 위한 인체공학적 디자인의 무선 키보드입니다. 긴 배터리 수명과 안정적인 블루투스 연결을 제공합니다.",
-    },
-    {
-      id: 23,
-      name: "무선 마우스",
-      price: 28000,
-      description:
-        "손목 피로를 줄여주는 인체공학적 디자인의 무선 마우스입니다. 정확한 센서와 긴 배터리 수명이 특징입니다.",
-    },
-    {
-      id: 24,
-      name: "모니터 받침대",
-      price: 32000,
-      description:
-        "목과 어깨의 피로를 줄여주는 높이 조절 가능한 모니터 받침대입니다. 견고한 스틸 소재로 제작되었습니다.",
-    },
-  ];
+  const response = await fetch("http://localhost:7070/api/v1/product/list");
+  if (!response.ok) {
+    console.log(response.status);
+    return;
+  }
+
+  const responseBody: ProductResponse = await response.json();
+  if (responseBody.statusCode != 200) {
+    alert("서버에 요청이 잘못되었습니다.");
+    return;
+  }
+
+  const products = responseBody.data;
 
   return <ClientPage products={products} />;
 }
